start deploy5

+++ db/migrate/20231022122245_change_nullable_constraints.rb
@@ -1,6 +1,6 @@
 class ChangeNullableConstraints < ActiveRecord::Migration[7.0]
   def change
-    change_column :payment_orders, :project_id, :integer, null: true
+    change_column :payment_orders, :project_id, :bigint, null: true
     change_column :payment_orders, :sci_id, :integer, null: true
     change_column :payment_orders, :spi_id, :integer, null: true
     change_column :payment_orders, :ballance_id, :integer, null: true
